chore: logical component separation
